resolve issues with render flags

+++ RayLua/interfaces/engineapi.c
@@ -51,7 +51,7 @@ void ForwardEngineUpdate(float dt) {
 
 static void InsertSprite(int sectid, float x, float y, float z) {
     long i = insspri_imp(sectid,x,y,z,mapref);
-    if (mapref->spri[i].tilnum > gmaltiles || mapref->spri[i].tilnum < 0)
+    if ((int)mapref->spri[i].tilnum > gmaltiles || mapref->spri[i].tilnum < 0)
         mapref->spri[i].tilnum = 1;
 }
